jusik_jong_be_recommendation/requirements.txt
pykrx>=1.0.45
finance-datareader>=0.9.50
yfinance>=0.2.38
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
pandas>=2.0.3
numpy>=1.24.0
rich>=13.7.0
python-dotenv>=1.0.0
pytz>=2024.1
tenacity>=8.2.3
html5lib>=1.1
anthropic
